fix: add coin api response interceptor
