+++ src/api/coinApi/fetchAllCurrentRates.ts
@@ -1,37 +1,20 @@
-import { AxiosError } from 'axios';
-
-import { coinApiAxiosInstance } from '@/api/coinApi/coinApiAxiosInstance';
+import { coinApiInstance } from '@/api/coinApi/config';
 import { GET_ALL_CURRENT_RATES_URL } from '@/constants/api/coinApi/endpoints';
 import { currencies } from '@/constants/currencies/currencies';
+import { CurrencyResponse } from '@/types/currencyResponse';
 
 export const fetchAllCurrentRates = async (baseCurrency: string = 'USD') => {
-  try {
-    const currencyCodes = currencies.map((currency) => currency.id).join(',');
+  const currencyCodes = currencies.map((currency) => currency.id).join(',');
 
-    const currentRates = await coinApiAxiosInstance.get(
-      `${GET_ALL_CURRENT_RATES_URL}/${baseCurrency}`,
-      {
-        params: {
-          filter_asset_id: currencyCodes,
-          invert: true,
-        },
+  const receivedCurrencies: CurrencyResponse = await coinApiInstance.get(
+    `${GET_ALL_CURRENT_RATES_URL}/${baseCurrency}`,
+    {
+      params: {
+        filter_asset_id: currencyCodes,
+        invert: true,
       },
-    );
+    },
+  );
 
-    return currentRates.data;
-  } catch (error) {
-    if (error instanceof AxiosError) {
-      if (error.response) {
-        console.log(error.response.data);
-        console.log(error.response.status);
-        console.log(error.response.headers);
-      } else if (error.request) {
-        console.log(error.request);
-      } else {
-        console.log('Error', error.message);
-      }
-      console.log(error.config);
-    }
-    return null;
-  }
+  return receivedCurrencies;
 };
