+++ src/services/currency/getCurrencies.ts
@@ -1,11 +1,11 @@
 import { fetchAllCurrentRates } from '@/api/coinApi/fetchAllCurrentRates';
 import { currencies } from '@/constants/currencies/currencies';
-import { CurrencyResponse } from '@/types/response';
+import { CurrencyResponseRate } from '@/types/currencyResponse';
 
 export const getCurrencies = async () => {
   const response = await fetchAllCurrentRates();
 
-  return response.rates.map((currencyWithRate: CurrencyResponse) => {
+  return response.rates.map((currencyWithRate: CurrencyResponseRate) => {
     const currency = currencies.find(
       (item) => item.id === currencyWithRate.asset_id_quote,
     );
