+++ src/types/currencyResponse.ts
@@ -1,0 +1,10 @@
+export interface CurrencyResponseRate {
+  time: string;
+  asset_id_quote: string;
+  rate: number;
+}
+
+export interface CurrencyResponse {
+  asset_id_base: string;
+  rates: CurrencyResponseRate[];
+}
